docs: removed comments

+++ client/store/cart.js
@@ -72,11 +72,7 @@ export const setCart = userId => async dispatch => {
 export const checkoutCart = (userId, cart) => async dispatch => {
   try {
     const response = await axios.post('/api/checkout', {userId, cart})
-    if (userId) {
-      // TODO: checkout for logged in users
-    } else {
-      window.localStorage.clear()
-    }
+    if (!userId) window.localStorage.clear()
     dispatch(clearCart())
   } catch (err) {
     console.error(err)
